requirements.txt
python-dotenv>=1.0.0
requests>=2.31.0
PyYAML>=6.0
beautifulsoup4>=4.12.0
fastapi>=0.104.0
uvicorn>=0.24.0
